broker/requirements.txt
python-dotenv==1.0.0
paho-mqtt==2.1.0